move clients
