+++ tests/Unit/Infrastructure/Http/ClientFactoryTest.php
@@ -3,9 +3,9 @@
 namespace ArtARTs36\MergeRequestLinter\Tests\Unit\Infrastructure\Http;
 
 use ArtARTs36\MergeRequestLinter\Configuration\HttpClientConfig;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Http\ClientFactory;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Http\MetricableClient;
-use ArtARTs36\MergeRequestLinter\Infrastructure\Http\NullClient;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Http\Client\ClientFactory;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Http\Client\MetricableClient;
+use ArtARTs36\MergeRequestLinter\Infrastructure\Http\Client\NullClient;
 use ArtARTs36\MergeRequestLinter\Report\Metrics\Manager\NullMetricManager;
 use ArtARTs36\MergeRequestLinter\Tests\TestCase;
 
@@ -20,7 +20,7 @@ final class ClientFactoryTest extends TestCase
     }
 
     /**
-     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Http\ClientFactory::create
+     * @covers \ArtARTs36\MergeRequestLinter\Infrastructure\Http\Client\ClientFactory::create
      * @param class-string $expectedClass
      * @dataProvider providerForTestCreate
      */
